python/requirements.txt
Flask[async]
discord.py
python-dotenv